requirements.txt
discord.py>=2.1.0
asyncio
googletrans
bs4
topggpy==2.0.0a
pushbullet.py
lxml
aiofiles